% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{methods}
\name{dbDataType,PqDriver,ANY-method}
\alias{dbDataType,PqConnection,ANY-method}
\alias{dbDataType,PqDriver,ANY-method}
\title{Determine database type for R vector.}
\usage{
\S4method{dbDataType}{PqDriver,ANY}(dbObj, obj)

\S4method{dbDataType}{PqConnection,ANY}(dbObj, obj)
}
\arguments{
\item{dbObj}{Postgres driver or connection.}

\item{obj}{Object to convert}
}
\description{
Determine database type for R vector.
}
\keyword{internal}

