% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{methods}
\name{dbDataType,PqDriver-method}
\alias{dbDataType,PqDriver-method}
\title{Determine database type for R vector.}
\usage{
\S4method{dbDataType}{PqDriver}(dbObj, obj)
}
\arguments{
\item{dbObj}{Postgres driver or connection.}

\item{obj}{Object to convert}
}
\description{
Determine database type for R vector.
}
\keyword{internal}
