% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbDataType_PqConnection.R
\name{dbDataType}
\alias{dbDataType_PqConnection}
\alias{dbDataType}
\alias{dbDataType_PqDriver}
\title{PostgreSQL database type}
\usage{
NULL
NULL
}

\arguments{
\item{dbObj}{Postgres driver or connection.}

\item{obj}{Object to convert}
}

\description{
PostgreSQL database type
}

\keyword{internal}
