% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/result.R
\docType{methods}
\name{dbFetch,PqResult-method}
\alias{dbClearResult,PqResult-method}
\alias{dbFetch,PqResult-method}
\alias{dbHasCompleted,PqResult-method}
\title{Fetch results into a data frame}
\usage{
\S4method{dbFetch}{PqResult}(res, n = -1, ...)

\S4method{dbHasCompleted}{PqResult}(res, ...)

\S4method{dbClearResult}{PqResult}(res, ...)
}
\arguments{
\item{res}{Code a \linkS4class{PqResult} produced by
\code{\link[DBI]{dbSendQuery}}.}

\item{n}{Number of rows to return. Currently must always return all data.}

\item{...}{Needed for compatibility with generic; currently ignored.}
}
\description{
Fetch results into a data frame
}

