% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connection.R
\docType{methods}
\name{dbGetException,PqConnection-method}
\alias{dbGetException,PqConnection-method}
\title{Get more details of error related to exception.}
\usage{
\S4method{dbGetException}{PqConnection}(conn, ...)
}
\arguments{
\item{conn}{A \code{PQConnection} object}

\item{...}{Needed for compatibility with generic; otherwise ignored.}
}
\description{
Get more details of error related to exception.
}
\examples{
con <- dbConnect(rpq::pq())
try(dbSendQuery(con, "BLAH"), silent = TRUE)
dbGetException(con)

try(dbSendQuery(con, "SELECT 1 + 'a'"), silent = TRUE)
dbGetException(con)
}

