% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/result.R
\docType{methods}
\name{dbSendQuery,PqConnection-method}
\alias{dbSendQuery,PqConnection-method}
\title{Send a query to the database.}
\usage{
\S4method{dbSendQuery}{PqConnection}(conn, statement, ...)
}
\arguments{
\item{conn}{A \code{\linkS4class{PqConnection}} created by \code{dbConnect}.}

\item{statement}{An SQL string to execture}

\item{...}{Another arguments needed for compatibility with generic (
currently ignored).}
}
\value{
On failure throws an error; otherwise returns an object
  of class \code{PqResult}.
}
\description{
Send a query to the database.
}
\examples{
con <- dbConnect(rpq::pq())
dbSendQuery(con, "SELECT * FROM pg_catalog.pg_tables")
}

