% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqConnection.R
\name{postgresIsTransacting}
\alias{postgresIsTransacting}
\title{Return whether a transaction is ongoing}
\usage{
postgresIsTransacting(conn)
}
\arguments{
\item{conn}{a \link[=PqConnection-class]{PqConnection} object, produced by \code{\link[DBI:dbConnect]{DBI::dbConnect()}}}
}
\value{
A boolean, indicating if a transaction is ongoing.
}
\description{
Detect whether the transaction is active for the given connection.
 A transaction might be started with \code{\link[DBI:dbBegin]{dbBegin()}} or wrapped within \code{\link[DBI:dbWithTransaction]{DBI::dbWithTransaction()}}.
}
