% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqConnection.R
\name{postgresWriteToLargeObject}
\alias{postgresWriteToLargeObject}
\title{Write to large object}
\usage{
postgresWriteToLargeObject(conn, oid, buf, bytes_to_write)
}
\arguments{
\item{conn}{a \linkS4class{PqConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{oid}{The oid to write to}

\item{buf}{Buffer of bytes to write. A vector of type "raw".}

\item{bytes_to_write}{Number of bytes to write from 'buf'}
}
\value{
The number of bytes written to large object
}
\description{
Writes a number of bytes from a raw vector to a large object.
}
\examples{
con         <- postgresDefault()
oid         <- dbGetQuery(con, "select lo_create(0) as oid")$oid[1]
raw_data    <- serialize(mtcars,NULL)
dbWithTransaction(con,{
    bytes_written <- postgresWriteToLargeObject(con, oid, raw_data, length(raw_data))
})
bytes       <- dbGetQuery(con, "select lo_get($1) as lo_data", params=list(oid))$lo_data[1]
mtcars_db   <- unserialize(unlist(bytes))
}
