% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbGetInfo_PqConnection.R
\name{PqConnection-class}
\alias{dbGetInfo_PqConnection}
\alias{dbGetInfo}
\alias{dbIsValid_PqConnection}
\alias{dbIsValid}
\alias{PqConnection}
\alias{format.PqConnection}
\alias{show_PqConnection}
\alias{show}
\title{PostgreSQL connection class}
\usage{
NULL
NULL
format.PqConnection(x, ...)
NULL
}

\description{
PostgreSQL connection class
}

\keyword{internal}
