% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbGetInfo_PqDriver.R
\name{PqDriver-class}
\alias{dbGetInfo_PqDriver}
\alias{dbGetInfo}
\alias{dbUnloadDriver_PqDriver}
\alias{dbUnloadDriver}
\alias{PqDriver}
\title{PostgreSQL driver class}
\usage{
NULL
NULL
}

\description{
PostgreSQL driver class
}

\keyword{internal}
