% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqGenerics.R
\name{pqExistsTable}
\alias{pqExistsTable}
\alias{pqExistsTable,PqConnection,character-method}
\title{Does a table exist (for the current user)?}
\usage{
pqExistsTable(conn, name, ...)

\S4method{pqExistsTable}{PqConnection,character}(conn, name, ...)
}
\arguments{
\item{conn}{a \linkS4class{PqConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{name}{a character string specifying a table name. Names will be
automatically quoted so you can use any sequence of characters, not
just any valid bare table name.}

\item{...}{Ignored.}
}
\description{
Returns whether a table or (materialized) view given by \code{name} is accessible
through this connection.
}
\examples{
# Examples only run on systems with a PostgreSQL connection:
run <- postgresHasDefault()

library(DBI)
if (run) con <- dbConnect(RPostgres::Postgres())
if (run) pqExistsTable(con, "mtcars")

if (run) dbWriteTable(con, "mtcars", mtcars, temporary = TRUE)
if (run) pqExistsTable(con, "mtcars")

if (run) dbDisconnect(con)

}
\seealso{
Other PqConnection generics: 
\code{\link{pqListFields}()},
\code{\link{pqListObjects}()},
\code{\link{pqListTables}()}
}
\concept{PqConnection generics}
