% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqGenerics.R
\name{pqListFields}
\alias{pqListFields}
\alias{pqListFields,DBIConnection,Id-method}
\title{List field names of a remote table}
\usage{
pqListFields(conn, name, ...)

\S4method{pqListFields}{DBIConnection,Id}(conn, name, ...)
}
\arguments{
\item{conn}{a \linkS4class{PqConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{name}{a character string specifying a table name. Names will be
automatically quoted so you can use any sequence of characters, not
just any valid bare table name.}

\item{...}{Ignored.}
}
\description{
List field names of a remote table
}
\examples{
# Examples only run on systems with a PostgreSQL connection:
run <- postgresHasDefault()

library(DBI)
if (run) con <- dbConnect(RPostgres::Postgres())

if (run) dbWriteTable(con, "mtcars", mtcars, temporary = TRUE)
if (run) pqListFields(con, "mtcars")

if (run) dbDisconnect(con)

}
\seealso{
\code{\link[=dbColumnInfo]{dbColumnInfo()}} to get the type of the fields.

Other PqConnection generics: 
\code{\link{pqExistsTable}()},
\code{\link{pqListObjects}()},
\code{\link{pqListTables}()}
}
\concept{PqConnection generics}
