% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqGenerics.R
\name{pqListObjects}
\alias{pqListObjects}
\alias{pqListObjects,PqConnection-method}
\title{List remote objects}
\usage{
pqListObjects(conn, prefix = NULL, ...)

\S4method{pqListObjects}{PqConnection}(conn, prefix = NULL, ...)
}
\arguments{
\item{conn}{a \linkS4class{PqConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{prefix}{A fully qualified path in the database's namespace, or \code{NULL}.
This argument will be processed with \code{\link[DBI:dbUnquoteIdentifier]{dbUnquoteIdentifier()}}.
If given the method will return all objects accessible through this prefix.}

\item{...}{Ignored.}
}
\description{
Returns the names of remote objects accessible through this connection as a
data frame.
This includes foreign and partitioned tables, (materialized) views, as well
as temporary tables.
Compared to \code{\link[=dbListTables]{dbListTables()}}, this method also enumerates tables and views
in schemas, and returns fully qualified identifiers to access these objects.
This allows exploration of all database objects available to the current
user, including those that can only be accessed by giving the full namespace.
}
\examples{
# Examples only run on systems with a PostgreSQL connection:
run <- postgresHasDefault()

library(DBI)
if (run) con <- dbConnect(RPostgres::Postgres())
if (run) pqListObjects(con)

if (run) dbWriteTable(con, "mtcars", mtcars, temporary = TRUE)
if (run) pqListObjects(con)

if (run) dbDisconnect(con)

}
\seealso{
Other PqConnection generics: 
\code{\link{pqExistsTable}()},
\code{\link{pqListFields}()},
\code{\link{pqListTables}()}
}
\concept{PqConnection generics}
