% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbColumnInfo_PqResult.R
\name{PqResult-class}
\alias{dbColumnInfo_PqResult}
\alias{dbColumnInfo}
\alias{dbGetRowCount_PqResult}
\alias{dbGetRowCount}
\alias{dbGetRowsAffected_PqResult}
\alias{dbGetRowsAffected}
\alias{dbGetStatement_PqResult}
\alias{dbGetStatement}
\alias{dbIsValid_PqDriver}
\alias{dbIsValid}
\alias{dbIsValid_PqResult}
\alias{PqResult}
\title{PostgreSQL result class}
\usage{
NULL
NULL
NULL
NULL
NULL
NULL
}

\description{
PostgreSQL result class
}

\keyword{internal}
