% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqConnection.R
\name{pqWaitForNotify}
\alias{pqWaitForNotify}
\title{Wait for and return any notifications that return within timeout}
\usage{
pqWaitForNotify(conn, timeout = 1)
}
\arguments{
\item{conn}{a \linkS4class{PqConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{timeout}{How long to wait, in seconds. Default 1}
}
\value{
If a notification was available, a list of:
\describe{
\item{channel}{Name of channel}
\item{pid}{PID of notifying server process}
\item{payload}{Content of notification}
}
If no notifications are available, return NULL
}
\description{
Once you subscribe to notifications with LISTEN, use this to wait for
responses on each channel.
}
\examples{
# For running the examples on systems without PostgreSQL connection:
if (postgresHasDefault()) {
  library(DBI)
  db <- dbConnect(RPostgres::Postgres())
  dbSendStatement(db, "LISTEN channel")
  # In another connection:-
  #     dbSendStatement(db2, "NOTIFY channel, 'hello'")
  n <- RPostgres::pqWaitForNotify(db)
  if (!is.null(n)) writeLines(c("Got a message:-", n$payload))
}
}
