% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPostgres-pkg.R
\docType{package}
\name{RPostgres-package}
\alias{RPostgres}
\alias{RPostgres-package}
\title{RPostgres: 'Rcpp' Interface to 'PostgreSQL'}
\description{
Fully 'DBI'-compliant 'Rcpp'-backed interface to 'PostgreSQL'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-dbi/RPostgres}
  \item Report bugs at \url{https://github.com/r-dbi/RPostgres/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{krlmlr+r@mailbox.org}

Authors:
\itemize{
  \item Hadley Wickham
  \item Jeroen Ooms
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
  \item R Consortium [copyright holder]
  \item Tomoaki Nishiyama (Code for encoding vectors into strings derived from RPostgreSQL) [contributor]
  \item Kungliga Tekniska Högskolan (Source code for timegm) [contributor]
}

}
