% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/connection.R
\docType{methods}
\name{dbConnect,PqDriver-method}
\alias{PqConnection-class}
\alias{dbConnect,PqDriver-method}
\title{Connect to a PostgreSQL database.}
\usage{
\S4method{dbConnect}{PqDriver}(drv, dbname = NULL, host = NULL,
  port = NULL, password = NULL, user = NULL, ...)
}
\arguments{
\item{drv}{\code{rpg::pq()}}

\item{dbname}{Database name. If \code{NULL}, defaults to the user name.}

\item{host,port}{Host and port. If \code{NULL}, will be retrieved from
\code{PGHOST} and \code{PGPORT} env vars.}

\item{user,password}{User name and password. If \code{NULLL}, will be
retrieved from \code{PGUSER} and \code{PGPASSWORD} envvars, or from the
appropriate line in \code{~/.pgpass}. See below for details.}

\item{...}{Other name-value pairs that describe additional connection
options as described at
\url{http://www.postgresql.org/docs/9.4/static/libpq-connect.html#LIBPQ-PARAMKEYWORDS}}
}
\description{
Connect to a PostgreSQL database.
}
\examples{
\donttest{
dbConnect(pq())
}
}

