% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqConnection.R
\docType{methods}
\name{dbDisconnect,PqConnection-method}
\alias{dbDisconnect,PqConnection-method}
\alias{dbConnect,PqDriver-method}
\title{Connect to a PostgreSQL database.}
\usage{
\S4method{dbDisconnect}{PqConnection}(conn, ...)

\S4method{dbConnect}{PqDriver}(drv, dbname = NULL, host = NULL,
  port = NULL, password = NULL, user = NULL, service = NULL, ...,
  bigint = c("integer64", "integer", "numeric", "character"))
}
\arguments{
\item{conn}{Connection to disconnect.}

\item{...}{Other name-value pairs that describe additional connection
options as described at
\url{http://www.postgresql.org/docs/9.4/static/libpq-connect.html#LIBPQ-PARAMKEYWORDS}}

\item{drv}{\code{RPostgres::Postgres()}}

\item{dbname}{Database name. If \code{NULL}, defaults to the user name.}

\item{host, port}{Host and port. If \code{NULL}, will be retrieved from
\code{PGHOST} and \code{PGPORT} env vars.}

\item{user, password}{User name and password. If \code{NULL}, will be
retrieved from \code{PGUSER} and \code{PGPASSWORD} envvars, or from the
appropriate line in \code{~/.pgpass}. See
\url{http://www.postgresql.org/docs/9.4/static/libpq-pgpass.html} for
more details.}

\item{service}{Name of service to connect as.  If \code{NULL}, will be
ignored.  Otherwise, connection parameters will be loaded from the pg_service.conf
file and used.  See \url{http://www.postgresql.org/docs/9.4/static/libpq-pgservice.html}
for details on this file and syntax.}
}
\description{
Note that manually disconnecting a connection is not necessary with RPostgres;
if you delete the object containing the connection, it will be automatcally
disconnected during the next GC.
}
\examples{
library(DBI)
con <- dbConnect(RPostgres::Postgres())
dbDisconnect(con)
}
