% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbDataType_PqConnection.R
\name{dbDataType}
\alias{dbDataType_PqConnection}
\alias{dbDataType,PqConnection-method}
\alias{dbDataType_PqDriver}
\alias{dbDataType,PqDriver-method}
\title{PostgreSQL database type}
\usage{
\S4method{dbDataType}{PqConnection}(dbObj, obj, ...)
\S4method{dbDataType}{PqDriver}(dbObj, obj, ...)
}

\arguments{
\item{dbObj}{Postgres driver or connection.}

\item{obj}{Object to convert}
}

\description{
PostgreSQL database type
}

\keyword{internal}
