% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbDataType_PqConnection.R,
%   R/dbDataType_PqDriver.R
\name{dbDataType_PqConnection}
\alias{dbDataType_PqConnection}
\alias{dbDataType,PqConnection-method}
\alias{dbDataType_PqDriver}
\alias{dbDataType,PqDriver-method}
\title{Determine database type for R vector.}
\usage{
\S4method{dbDataType}{PqConnection}(dbObj, obj, ...)

\S4method{dbDataType}{PqDriver}(dbObj, obj, ...)
}
\arguments{
\item{dbObj}{Postgres driver or connection.}

\item{obj}{Object to convert}
}
\description{
Determine database type for R vector.
}
\keyword{internal}
