% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbBind_PqResult.R
\name{postgres-query}
\alias{dbBind_PqResult}
\alias{dbBind,PqResult-method}
\alias{dbClearResult_PqResult}
\alias{dbClearResult,PqResult-method}
\alias{dbFetch_PqResult}
\alias{dbFetch,PqResult-method}
\alias{dbHasCompleted_PqResult}
\alias{dbHasCompleted,PqResult-method}
\alias{dbSendQuery_PqConnection}
\alias{dbSendQuery,PqConnection-method}
\title{PostgreSQL query}
\usage{
\S4method{dbBind}{PqResult}(res, params, ...)
\S4method{dbClearResult}{PqResult}(res, ...)
\S4method{dbFetch}{PqResult}(res, n = -1, ..., row.names = FALSE)
\S4method{dbHasCompleted}{PqResult}(res, ...)
\S4method{dbSendQuery}{PqConnection}(conn, statement, params = NULL, ..., immediate = FALSE)
}

\arguments{
\item{res}{Code a [PqResult-class] produced by
  [DBI::dbSendQuery()].}

\item{n}{Number of rows to return. If less than zero returns all rows.}

\item{row.names}{Whether to convert row names to a column. Use `FALSE`
  to discard row names.}

\item{immediate}{If `TRUE`, uses the `PGsendQuery()` API instead of `PGprepare()`.
  This allows to pass multiple statements and turns off the ability to pass parameters.}

\item{conn}{A [PqConnection-class] created by [dbConnect()].}

\item{statement}{An SQL string to execute.}

\item{params}{A list of query parameters to be substituted into
  a parameterised query. Query parameters are sent as strings, and the
  correct type is imputed by PostgreSQL. If this fails, you can manually
  cast the parameter with e.g. `"$1::bigint"`.}

\item{...}{Other arguments needed for compatibility with generic (currently
  ignored).}
}

\description{
PostgreSQL query
}

