% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbBind_PqResult.R
\name{postgres-query}
\alias{dbBind_PqResult}
\alias{dbBind}
\alias{dbClearResult_PqResult}
\alias{dbClearResult}
\alias{dbFetch_PqResult}
\alias{dbFetch}
\alias{dbHasCompleted_PqResult}
\alias{dbHasCompleted}
\alias{dbSendQuery_PqConnection}
\alias{dbSendQuery}
\title{PostgreSQL query}
\usage{
NULL
NULL
NULL
NULL
NULL
}

\arguments{
\item{res}{Code a [PqResult-class] produced by
  [DBI::dbSendQuery()].}

\item{n}{Number of rows to return. If less than zero returns all rows.}

\item{immediate}{If `TRUE`, uses the `PGsendQuery()` API instead of `PGprepare()`.
  This allows to pass multiple statements and turns off the ability to pass parameters.}

\item{conn}{A [PqConnection-class] created by [dbConnect()].}

\item{statement}{An SQL string to execute.}

\item{params}{A list of query parameters to be substituted into
  a parameterised query. Query parameters are sent as strings, and the
  correct type is imputed by PostgreSQL. If this fails, you can manually
  cast the parameter with e.g. `"$1::bigint"`.}

\item{...}{Other arguments needed for compatibility with generic (currently
  ignored).}
}

\description{
PostgreSQL query
}

