% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbAppendTable_PqConnection.R
\name{postgres-tables}
\alias{dbAppendTable_PqConnection}
\alias{dbAppendTable,PqConnection-method}
\alias{dbExistsTable_PqConnection}
\alias{dbExistsTable,PqConnection,character-method}
\alias{dbExistsTable,PqConnection,Id-method}
\alias{dbListFields_PqConnection}
\alias{dbListFields,PqConnection,character-method}
\alias{dbListFields,PqConnection,Id-method}
\alias{dbListObjects_PqConnection_ANY}
\alias{dbListObjects,PqConnection-method}
\alias{dbListTables_PqConnection}
\alias{dbListTables,PqConnection-method}
\alias{dbReadTable_PqConnection_character}
\alias{dbReadTable,PqConnection,character-method}
\alias{dbReadTable,PqConnection,Id-method}
\alias{dbRemoveTable_PqConnection_character}
\alias{dbRemoveTable,PqConnection,character-method}
\alias{dbRemoveTable,PqConnection,Id-method}
\alias{dbWriteTable_PqConnection_character_data.frame}
\alias{dbWriteTable,PqConnection,character-method}
\alias{dbWriteTable,PqConnection,Id-method}
\alias{sqlData_PqConnection}
\alias{sqlData,PqConnection-method}
\title{PostgreSQL tables}
\usage{
\S4method{dbAppendTable}{PqConnection}(conn, name, value, copy = NULL, ..., row.names = NULL)
\S4method{dbExistsTable}{PqConnection,character}(conn, name, ...)
\S4method{dbListFields}{PqConnection,character}(conn, name, ...)
\S4method{dbListObjects}{PqConnection}(conn, prefix = NULL, ...)
\S4method{dbListTables}{PqConnection}(conn, ...)
\S4method{dbReadTable}{PqConnection,character}(conn, name, ..., check.names = TRUE, row.names = FALSE)
\S4method{dbRemoveTable}{PqConnection,character}(conn, name, ..., temporary = FALSE, fail_if_missing = TRUE)
\S4method{dbWriteTable}{PqConnection,character}(
  conn, name, value, ...,
  row.names = FALSE, overwrite = FALSE, append = FALSE,
  field.types = NULL, temporary = FALSE, copy = NULL
)
NULL
}

\arguments{
\item{row.names}{Must be `NULL`.}

\item{prefix}{A [DBI::Id-class] object identifying a schema to list
  objects within, or `NULL` to list all tables and schemas.}

\item{check.names}{If `TRUE`, the default, column names will be
  converted to valid R identifiers.}

\item{temporary}{If `TRUE`, only temporary tables are considered.}

\item{fail_if_missing}{If `FALSE`, `dbRemoveTable()` succeeds if the
  table doesn't exist.}

\item{...}{Ignored.}

\item{conn}{a [PqConnection-class] object, produced by
  [DBI::dbConnect()]}

\item{name}{a character string specifying a table name. Names will be
  automatically quoted so you can use any sequence of characters, not
  just any valid bare table name.
  Alternatively, pass a name quoted with [dbQuoteIdentifier()],
  an [Id()] object, or a string escaped with [SQL()].}

\item{value}{A data.frame to write to the database.}

\item{overwrite}{a logical specifying whether to overwrite an existing table
  or not. Its default is `FALSE`.}

\item{append}{a logical specifying whether to append to an existing table
  in the DBMS. Its default is `FALSE`.}

\item{field.types}{character vector of named SQL field types where
  the names are the names of new table's columns.
  If missing, types are inferred with [DBI::dbDataType()]).
  The types can only be specified with `append = FALSE`.}

\item{copy}{If `TRUE`, serializes the data frame to a single string
  and uses `COPY name FROM stdin`, which is often fast. If `FALSE`, generates
  a single SQL string that is but widely supported.
  The default maps to `TRUE`.}
}

\description{
[dbAppendTable()] is overridden because \pkg{rpsql}
uses placeholders of the form `$1`, `$2` etc. instead of `?`.
}

