% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tables.R
\docType{methods}
\name{postgres-tables}
\alias{dbExistsTable,PqConnection-method}
\alias{dbListTables,PqConnection-method}
\alias{dbReadTable,PqConnection,character-method}
\alias{dbRemoveTable,PqConnection,character-method}
\alias{dbWriteTable,PqConnection,character,data.frame-method}
\alias{postgres-tables}
\title{Convenience functions for reading/writing DBMS tables}
\usage{
\S4method{dbWriteTable}{PqConnection,character,data.frame}(conn, name, value,
  row.names = NA, overwrite = FALSE, append = FALSE, field.types = NULL,
  temporary = FALSE)

\S4method{dbReadTable}{PqConnection,character}(conn, name, row.names = NA)

\S4method{dbListTables}{PqConnection}(conn)

\S4method{dbExistsTable}{PqConnection}(conn, name)

\S4method{dbRemoveTable}{PqConnection,character}(conn, name)
}
\arguments{
\item{conn}{a \code{\linkS4class{PqConnection}} object, produced by
\code{\link[DBI]{dbConnect}}}

\item{name}{a character string specifying a table name. Names will be
automatically quoted so you can use any sequence of characaters, not
just any valid bare table name.}

\item{value}{A data.frame to write to the database.}

\item{row.names}{Either \code{TRUE}, \code{FALSE}, \code{NA} or a string.

  If \code{TRUE}, always translate row names to a column called "row_names".
  If \code{FALSE}, never translate row names. If \code{NA}, translate
  rownames only if they're a character vector.

  A string is equivalent to \code{TRUE}, but allows you to override the
  default name.

  For backward compatibility, \code{NULL} is equivalent to \code{FALSE}.}

\item{overwrite}{a logical specifying whether to overwrite an existing table
or not. Its default is \code{FALSE}.}

\item{append}{a logical specifying whether to append to an existing table
in the DBMS. Its default is \code{FALSE}.}

\item{field.types}{character vector of named SQL field types where
the names are the names of new table's columns. If missing, types inferred
with \code{\link[DBI]{dbDataType}}).}

\item{temporary}{If \code{TRUE}, will generate a temporary table statement.}
}
\description{
Convenience functions for reading/writing DBMS tables
}
\examples{
con <- dbConnect(RPostgres::Postgres())
dbListTables(con)
dbWriteTable(con, "mtcars", mtcars, temporary = TRUE)
dbReadTable(con, "mtcars")

dbListTables(con)
dbExistsTable(con, "mtcars")

# A zero row data frame just creates a table definition.
dbWriteTable(con, "mtcars2", mtcars[0, ], temporary = TRUE)
dbReadTable(con, "mtcars2")

dbDisconnect(con)
}

