% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbBegin_PqConnection.R
\name{postgres-transactions}
\alias{dbBegin_PqConnection}
\alias{dbBegin}
\alias{dbCommit_PqConnection}
\alias{dbCommit}
\alias{dbRollback_PqConnection}
\alias{dbRollback}
\title{PostgreSQL transations}
\usage{
NULL
NULL
NULL
}

\arguments{
\item{name}{If provided, uses the `SAVEPOINT` SQL syntax
  to establish, remove (commit) or undo a ßsavepoint.}

\item{conn}{a [PqConnection-class] object, produced by
  [DBI::dbConnect()]}

\item{...}{Unused, for extensibility.}
}

\value{
A boolean, indicating success or failure.
}

\description{
PostgreSQL transations
}

