% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbBegin_PqConnection.R
\name{postgres-transactions}
\alias{dbBegin_PqConnection}
\alias{dbBegin,PqConnection-method}
\alias{dbCommit_PqConnection}
\alias{dbCommit,PqConnection-method}
\alias{dbRollback_PqConnection}
\alias{dbRollback,PqConnection-method}
\title{PostgreSQL transations}
\usage{
\S4method{dbBegin}{PqConnection}(conn, ..., name = NULL)
\S4method{dbCommit}{PqConnection}(conn, ..., name = NULL)
\S4method{dbRollback}{PqConnection}(conn, ..., name = NULL)
}

\arguments{
\item{name}{If provided, uses the `SAVEPOINT` SQL syntax
  to establish, remove (commit) or undo a ßsavepoint.}

\item{conn}{a [PqConnection-class] object, produced by
  [DBI::dbConnect()]}

\item{...}{Unused, for extensibility.}
}

\value{
A boolean, indicating success or failure.
}

\description{
PostgreSQL transations
}

