% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbConnect_PqDriver.R
\name{Postgres}
\alias{dbConnect_PqDriver}
\alias{dbConnect}
\alias{dbDisconnect_PqConnection}
\alias{dbDisconnect}
\alias{Postgres}
\title{Connect to a PostgreSQL database}
\usage{
NULL
NULL
Postgres()
}

\arguments{
\item{drv}{[DBI::DBIDriver-class]. Use [Postgres()] to connect to a
  PostgreSQL database. Use an existing [DBI::DBIConnection-class] object to
  clone an existing connection.}

\item{dbname}{Database name. If `NULL`, defaults to the user name.
  Note that this argument can only contain the database name, it will not
  be parsed as a connection string (internally, `expand_dbname` is set to
  `false` in the call to
  [`PQconnectdbParams()`](https://www.postgresql.org/docs/current/libpq-connect.html)).}

\item{user}{User name and password. If `NULL`, will be
  retrieved from `PGUSER` and `PGPASSWORD` envvars, or from the
  appropriate line in `~/.pgpass`. See
  <https://www.postgresql.org/docs/current/libpq-pgpass.html> for
  more details.}

\item{password}{User name and password. If `NULL`, will be
  retrieved from `PGUSER` and `PGPASSWORD` envvars, or from the
  appropriate line in `~/.pgpass`. See
  <https://www.postgresql.org/docs/current/libpq-pgpass.html> for
  more details.}

\item{host}{Host and port. If `NULL`, will be retrieved from
  `PGHOST` and `PGPORT` env vars.}

\item{port}{Host and port. If `NULL`, will be retrieved from
  `PGHOST` and `PGPORT` env vars.}

\item{service}{Name of service to connect as.  If `NULL`, will be
  ignored.  Otherwise, connection parameters will be loaded from the pg_service.conf
  file and used.  See <https://www.postgresql.org/docs/current/libpq-pgservice.html>
  for details on this file and syntax.}

\item{...}{Other name-value pairs that describe additional connection
  options as described at
  <https://www.postgresql.org/docs/current/libpq-connect.html#LIBPQ-PARAMKEYWORDS>}

\item{bigint}{The R type that 64-bit integer types should be mapped to,
  default is [bit64::integer64], which allows the full range of 64 bit
  integers.}

\item{check_interrupts}{Should user interrupts be checked during the query execution (before
  first row of data is available)? Setting to `TRUE` allows interruption of queries
  running too long.}

\item{timezone}{Sets the timezone for the connection. The default is `"UTC"`.
  If `NULL` then no timezone is set, which defaults to the server's time zone.}

\item{timezone_out}{The time zone returned to R, defaults to `timezone`.
  If you want to display datetime values in the local timezone,
  set to [Sys.timezone()] or `""`.
  This setting does not change the time values returned, only their display.}

\item{conn}{Connection to disconnect.}
}

\description{
`DBI::dbConnect()` establishes a connection to a database.
Set `drv = Postgres()` to connect to a PostgreSQL database

Manually disconnecting a connection is not necessary with \pkg{rpsql},
but still recommended;
if you delete the object containing the connection, it will be automatically
disconnected during the next GC with a warning.
}

