% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/driver.R
\name{Postgres}
\alias{Postgres}
\title{Postgres driver}
\usage{
Postgres()
}
\description{
This driver never needs to be unloaded and hence \code{dbUnload()} is a
null-op.
}
\examples{
RPostgres::Postgres()
}

