% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqConnection.R
\name{postgresExportLargeObject}
\alias{postgresExportLargeObject}
\title{Exports a large object to file}
\usage{
postgresExportLargeObject(conn, oid, filepath = NULL)
}
\arguments{
\item{conn}{a \linkS4class{PqConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{oid}{the object identifier (Oid) of the large object to export}

\item{filepath}{a path where the large object should be exported}
}
\value{
1 on success (invisibly), or stops with an error
}
\description{
Exports a large object from the database to a file on disk
}
\examples{
\dontrun{
con <- postgresDefault()
filepath <- 'your_image.png'
dbWithTransaction(con, {
  oid <- postgresImportLargeObject(con, filepath)
})
# Later, export the large object back to a file
dbWithTransaction(con, {
  postgresExportLargeObject(con, oid, 'exported_image.png')
})
}
}