% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in default.R
\name{postgresHasDefault}
\alias{postgresHasDefault}
\alias{postgresDefault}
\title{Check if default database is available.}
\usage{
postgresHasDefault(...)
postgresDefault(...)
}

\arguments{
\item{...}{Additional arguments passed on to [dbConnect()]}
}

\description{
rpsql examples and tests connect to a default database via
`dbConnect(`[rpsql::Postgres()]`)`. This function checks if that
database is available, and if not, displays an informative message.
}

\examples{
if (postgresHasDefault()) {
  db <- postgresDefault()
  print(dbListTables(db))
  dbDisconnect(db)
} else {
  message("No database connection.")
}
}

