% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqConnection.R
\name{postgresImportLargeObject}
\alias{postgresImportLargeObject}
\title{Imports a large object from file}
\usage{
postgresImportLargeObject(conn, filepath = NULL, oid = 0)
}
\arguments{
\item{conn}{a \linkS4class{PqConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}}

\item{filepath}{A path to the large object to import}

\item{oid}{The oid to write to. Defaults to 0 which assigns an unused oid}
}
\value{
An integer, the identifier of the large object
}
\description{
Returns an object idenfier (Oid) for the imported large object
}
\examples{
con       <- postgresDefault()
path_to_file <- 'my_image.png'
dbWithTransaction(con, { 
 oid <- postgresImportLargeObject(con, test_file_path)
})
}
