% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in PqConnection.R
\name{postgresIsTransacting}
\alias{postgresIsTransacting}
\title{Return whether a transaction is ongoing}
\usage{
postgresIsTransacting(conn)
}

\arguments{
\item{conn}{a [PqConnection-class] object, produced by
  [DBI::dbConnect()]}
}

\value{
A boolean, indicating if a transaction is ongoing.
}

\description{
Detect whether the transaction is active for the given connection. A
transaction might be started with [dbBegin()] or wrapped within
[DBI::dbWithTransaction()].
}

