% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in PqConnection.R
\name{postgresWaitForNotify}
\alias{postgresWaitForNotify}
\title{Wait for and return any notifications that return within timeout}
\usage{
postgresWaitForNotify(conn, timeout = 1)
}

\arguments{
\item{conn}{a [PqConnection-class] object, produced by
  [DBI::dbConnect()]}

\item{timeout}{How long to wait, in seconds. Default 1}
}

\value{
If a notification was available, a list of:
\describe{
  \item{channel}{Name of channel}
  \item{pid}{PID of notifying server process}
  \item{payload}{Content of notification}
}
If no notifications are available, return NULL
}

\description{
Once you subscribe to notifications with LISTEN, use this to wait for
responses on each channel.
}

