% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbGetInfo_PqConnection.R
\name{PqConnection-class}
\alias{dbGetInfo_PqConnection}
\alias{dbGetInfo,PqConnection-method}
\alias{dbIsValid_PqConnection}
\alias{dbIsValid,PqConnection-method}
\alias{PqConnection}
\alias{format.PqConnection}
\alias{show_PqConnection}
\alias{show,PqConnection-method}
\title{PostgreSQL connection class}
\usage{
NULL
NULL
\method{format}{PqConnection}(x, ...)
NULL
}

\description{
PostgreSQL connection class
}

\keyword{internal}
