% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqConnection.R, R/dbGetInfo_PqConnection.R,
%   R/dbIsValid_PqConnection.R, R/show_PqConnection.R
\docType{class}
\name{PqConnection-class}
\alias{PqConnection-class}
\alias{format.PqConnection}
\alias{dbGetInfo_PqConnection}
\alias{dbGetInfo,PqConnection-method}
\alias{dbIsValid_PqConnection}
\alias{dbIsValid,PqConnection-method}
\alias{show_PqConnection}
\alias{show,PqConnection-method}
\title{PqConnection and methods.}
\usage{
\method{format}{PqConnection}(x, ...)

\S4method{dbGetInfo}{PqConnection}(dbObj, ...)

\S4method{dbIsValid}{PqConnection}(dbObj, ...)

\S4method{show}{PqConnection}(object)
}
\description{
PqConnection and methods.
}
\keyword{internal}
