% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqConnection.R
\docType{class}
\name{PqConnection-class}
\alias{PqConnection-class}
\alias{format.PqConnection}
\alias{show,PqConnection-method}
\alias{dbIsValid,PqConnection-method}
\alias{dbGetInfo,PqConnection-method}
\title{PqConnection and methods.}
\usage{
\method{format}{PqConnection}(x, ...)

\S4method{show}{PqConnection}(object)

\S4method{dbIsValid}{PqConnection}(dbObj, ...)

\S4method{dbGetInfo}{PqConnection}(dbObj, ...)
}
\description{
PqConnection and methods.
}
\keyword{internal}
