% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\docType{class}
\name{PqConnection-class}
\alias{PqConnection-class}
\alias{dbGetInfo,PqConnection-method}
\alias{show,PqConnection-method}
\title{PqConnection and methods.}
\usage{
\S4method{dbGetInfo}{PqConnection}(dbObj, ...)

\S4method{show}{PqConnection}(object)
}
\description{
PqConnection and methods.
}
\keyword{internal}
