% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbGetInfo_PqDriver.R
\name{PqDriver-class}
\alias{dbGetInfo_PqDriver}
\alias{dbGetInfo,PqDriver-method}
\alias{dbIsValid_PqDriver}
\alias{dbIsValid,PqDriver-method}
\alias{dbUnloadDriver_PqDriver}
\alias{dbUnloadDriver,PqDriver-method}
\alias{PqDriver}
\title{PostgreSQL driver class}
\usage{
NULL
NULL
NULL
}

\description{
PostgreSQL driver class
}

\keyword{internal}
