% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqGenerics.R
\name{pqListTables}
\alias{pqListTables}
\alias{pqListTables,PqConnection-method}
\title{List remote tables}
\usage{
pqListTables(conn, ...)

\S4method{pqListTables}{PqConnection}(conn, ...)
}
\arguments{
\item{conn}{a \linkS4class{PqConnection} object, produced by
\code{\link[DBI:dbConnect]{DBI::dbConnect()}}.}

\item{...}{Ignored.}
}
\description{
Returns the unquoted names of remote tables and table-like objects accessible
through this connection.
This includes foreign and partitioned tables, (materialized) views, as well
as temporary objects.
}
\examples{
# Examples only run on systems with a PostgreSQL connection:
run <- postgresHasDefault()

library(DBI)
if (run) con <- dbConnect(RPostgres::Postgres())
if (run) pqListTables(con)

if (run) dbWriteTable(con, "mtcars", mtcars, temporary = TRUE)
if (run) pqListTables(con)

if (run) dbDisconnect(con)

}
\seealso{
Other PqConnection generics: 
\code{\link{pqExistsTable}()},
\code{\link{pqListFields}()},
\code{\link{pqListObjects}()}
}
\concept{PqConnection generics}
