% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbColumnInfo_PqResult.R
\name{PqResult-class}
\alias{dbColumnInfo_PqResult}
\alias{dbColumnInfo,PqResult-method}
\alias{dbGetRowCount_PqResult}
\alias{dbGetRowCount,PqResult-method}
\alias{dbGetRowsAffected_PqResult}
\alias{dbGetRowsAffected,PqResult-method}
\alias{dbGetStatement_PqResult}
\alias{dbGetStatement,PqResult-method}
\alias{dbIsValid_PqResult}
\alias{dbIsValid,PqResult-method}
\alias{PqResult}
\title{PostgreSQL result class}
\usage{
NULL
NULL
NULL
NULL
NULL
}

\description{
PostgreSQL result class
}

\keyword{internal}
