% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PqResult.R, R/dbColumnInfo_PqResult.R,
%   R/dbGetRowCount_PqResult.R, R/dbGetRowsAffected_PqResult.R,
%   R/dbGetStatement_PqResult.R, R/dbIsValid_PqDriver.R, R/dbIsValid_PqResult.R
\docType{class}
\name{PqResult-class}
\alias{PqResult-class}
\alias{dbColumnInfo_PqResult}
\alias{dbColumnInfo,PqResult-method}
\alias{dbGetRowCount_PqResult}
\alias{dbGetRowCount,PqResult-method}
\alias{dbGetRowsAffected_PqResult}
\alias{dbGetRowsAffected,PqResult-method}
\alias{dbGetStatement_PqResult}
\alias{dbGetStatement,PqResult-method}
\alias{dbIsValid_PqDriver}
\alias{dbIsValid,PqDriver-method}
\alias{dbIsValid_PqResult}
\alias{dbIsValid,PqResult-method}
\title{PostgreSQL results.}
\usage{
\S4method{dbColumnInfo}{PqResult}(res, ...)

\S4method{dbGetRowCount}{PqResult}(res, ...)

\S4method{dbGetRowsAffected}{PqResult}(res, ...)

\S4method{dbGetStatement}{PqResult}(res, ...)

\S4method{dbIsValid}{PqDriver}(dbObj, ...)

\S4method{dbIsValid}{PqResult}(dbObj, ...)
}
\description{
PostgreSQL results.
}
\keyword{internal}
