% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/result.R
\docType{class}
\name{PqResult-class}
\alias{PqResult-class}
\alias{dbColumnInfo,PqResult-method}
\alias{dbGetRowCount,PqResult-method}
\alias{dbGetStatement,PqResult-method}
\alias{dbIsValid,PqResult-method}
\alias{show,PqResult-method}
\title{PostgreSQL results.}
\usage{
\S4method{dbGetStatement}{PqResult}(res, ...)

\S4method{dbIsValid}{PqResult}(dbObj, ...)

\S4method{dbGetRowCount}{PqResult}(res, ...)

\S4method{dbColumnInfo}{PqResult}(res, ...)

\S4method{show}{PqResult}(object)
}
\description{
PostgreSQL results.
}
\keyword{internal}

