% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbQuoteIdentifier_PqConnection_character.R
\name{quote}
\alias{dbQuoteIdentifier_PqConnection}
\alias{dbQuoteIdentifier,PqConnection,character-method}
\alias{dbQuoteIdentifier,PqConnection,SQL-method}
\alias{dbQuoteIdentifier,PqConnection,Id-method}
\alias{dbQuoteLiteral_PqConnection}
\alias{dbQuoteLiteral,PqConnection-method}
\alias{dbQuoteString_PqConnection}
\alias{dbQuoteString,PqConnection,character-method}
\alias{dbQuoteString,PqConnection,SQL-method}
\alias{dbUnquoteIdentifier_PqConnection}
\alias{dbUnquoteIdentifier,PqConnection-method}
\title{PostgreSQL quote}
\usage{
\S4method{dbQuoteIdentifier}{PqConnection,character}(conn, x, ...)
\S4method{dbQuoteLiteral}{PqConnection}(conn, x, ...)
\S4method{dbQuoteString}{PqConnection,character}(conn, x, ...)
\S4method{dbUnquoteIdentifier}{PqConnection}(conn, x, ...)
}

\arguments{
\item{conn}{A [PqConnection-class] created by `dbConnect()`}

\item{x}{A character vector to be quoted.}

\item{...}{Other arguments needed for compatibility with generic (currently
  ignored).}
}

\description{
PostgreSQL quote
}

