% Generated by tinyroxygen: do not edit by hand
% Please edit documentation in dbQuoteIdentifier_PqConnection_character.R
\name{quote}
\alias{dbQuoteIdentifier_PqConnection}
\alias{dbQuoteIdentifier}
\alias{dbQuoteLiteral_PqConnection}
\alias{dbQuoteLiteral}
\alias{dbQuoteString_PqConnection}
\alias{dbQuoteString}
\alias{dbUnquoteIdentifier_PqConnection}
\alias{dbUnquoteIdentifier}
\title{PostgreSQL quote}
\usage{
NULL
NULL
NULL
NULL
}

\arguments{
\item{conn}{A [PqConnection-class] created by `dbConnect()`}

\item{x}{A character vector to be quoted.}

\item{...}{Other arguments needed for compatibility with generic (currently
  ignored).}
}

\description{
PostgreSQL quote
}

