% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quote.R
\docType{methods}
\name{quote}
\alias{dbQuoteIdentifier,PqConnection,ANY-method}
\alias{dbQuoteString,PqConnection,ANY-method}
\alias{quote}
\title{Quote postgres strings and identifiers}
\usage{
\S4method{dbQuoteString}{PqConnection,ANY}(conn, x, ...)

\S4method{dbQuoteIdentifier}{PqConnection,ANY}(conn, x, ...)
}
\description{
Quote postgres strings and identifiers
}

