% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quote.R
\docType{methods}
\name{quote}
\alias{quote}
\alias{dbQuoteString,PqConnection,character-method}
\alias{dbQuoteString,PqConnection,SQL-method}
\alias{dbQuoteIdentifier,PqConnection,character-method}
\alias{dbQuoteIdentifier,PqConnection,SQL-method}
\alias{dbQuoteLiteral,PqConnection,logical-method}
\alias{dbQuoteLiteral,PqConnection,integer-method}
\alias{dbQuoteLiteral,PqConnection,numeric-method}
\alias{dbQuoteLiteral,PqConnection,factor-method}
\alias{dbQuoteLiteral,PqConnection,Date-method}
\alias{dbQuoteLiteral,PqConnection,POSIXt-method}
\alias{dbQuoteLiteral,PqConnection,difftime-method}
\alias{dbQuoteLiteral,PqConnection,list-method}
\title{Quote postgres strings and identifiers.}
\usage{
\S4method{dbQuoteString}{PqConnection,character}(conn, x, ...)

\S4method{dbQuoteString}{PqConnection,SQL}(conn, x, ...)

\S4method{dbQuoteIdentifier}{PqConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{PqConnection,SQL}(conn, x, ...)

\S4method{dbQuoteLiteral}{PqConnection,logical}(conn, x, ...)

\S4method{dbQuoteLiteral}{PqConnection,integer}(conn, x, ...)

\S4method{dbQuoteLiteral}{PqConnection,numeric}(conn, x, ...)

\S4method{dbQuoteLiteral}{PqConnection,factor}(conn, x, ...)

\S4method{dbQuoteLiteral}{PqConnection,Date}(conn, x, ...)

\S4method{dbQuoteLiteral}{PqConnection,POSIXt}(conn, x, ...)

\S4method{dbQuoteLiteral}{PqConnection,difftime}(conn, x, ...)

\S4method{dbQuoteLiteral}{PqConnection,list}(conn, x, ...)
}
\arguments{
\item{conn}{A \linkS4class{PqConnection} created by \code{dbConnect()}}

\item{x}{A character to escaped}

\item{...}{Other arguments needed for compatibility with generic}
}
\description{
Quote postgres strings and identifiers.
}
\examples{
library(DBI)
con <- dbConnect(RPostgres::Postgres())

x <- c("a", "b c", "d'e", "\\\\f")
dbQuoteString(con, x)
dbQuoteIdentifier(con, x)
}
