% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quote.R
\docType{methods}
\name{quote}
\alias{dbQuoteIdentifier,PqConnection,character-method}
\alias{dbQuoteString,PqConnection,character-method}
\alias{quote}
\title{Quote postgres strings and identifiers.}
\usage{
\S4method{dbQuoteString}{PqConnection,character}(conn, x, ...)

\S4method{dbQuoteIdentifier}{PqConnection,character}(conn, x, ...)
}
\arguments{
\item{conn}{A \linkS4class{PqConnection} created by \code{dbConnect()}}

\item{x}{A character to escaped}

\item{...}{Other arguments needed for compatibility with generic}
}
\description{
Quote postgres strings and identifiers.
}
\examples{
library(DBI)
con <- dbConnect(RPostgres::Postgres())

x <- c("a", "b c", "d'e", "\\\\f")
dbQuoteString(con, x)
dbQuoteIdentifier(con, x)
}

