% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Redshift.R
\docType{class}
\name{Redshift}
\alias{Redshift}
\alias{RedshiftDriver-class}
\alias{RedshiftConnection-class}
\alias{dbConnect,RedshiftDriver-method}
\title{Redshift driver/connection}
\usage{
Redshift()

\S4method{dbConnect}{RedshiftDriver}(
  drv,
  dbname = NULL,
  host = NULL,
  port = NULL,
  password = NULL,
  user = NULL,
  service = NULL,
  ...,
  bigint = c("integer64", "integer", "numeric", "character"),
  check_interrupts = FALSE,
  timezone = "UTC"
)
}
\description{
Redshift currently uses all the same method as Postgres, but allows
provides an extension point for future methods and downstream packages.
}
