% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPostgres-pkg.R
\docType{package}
\name{RPostgres-package}
\alias{RPostgres}
\alias{RPostgres-package}
\title{RPostgres: Experimental Rcpp Interface to PostgreSQL}
\description{
DBI-compatible Rcpp-backed interface to PostgreSQL.
}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@rstudio.com}

Authors:
\itemize{
  \item Jeroen Ooms
}

Other contributors:
\itemize{
  \item Kirill Müller [contributor]
  \item RStudio [copyright holder]
  \item Tomoaki Nishiyama (Code for encoding vectors into strings derived from RPostgreSQL) [contributor]
  \item Kungliga Tekniska Högskolan (Source code for timegm) [contributor]
}

}
