% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPostgres-pkg.R
\docType{package}
\name{RPostgres-package}
\alias{RPostgres}
\alias{RPostgres-package}
\title{RPostgres: Rcpp Interface to PostgreSQL}
\description{
Fully DBI-compliant Rcpp-backed interface to PostgreSQL \url{https://www.postgresql.org/}, an open-source relational database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rpostgres.r-dbi.org}
  \item \url{https://github.com/r-dbi/RPostgres}
  \item Report bugs at \url{https://github.com/r-dbi/RPostgres/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

Authors:
\itemize{
  \item Hadley Wickham
  \item Jeroen Ooms
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
  \item R Consortium [funder]
  \item Tomoaki Nishiyama (Code for encoding vectors into strings derived from RPostgreSQL) [contributor]
}

}
